feat: enhance testing setup with new e2e configurations
