+++ vite.e2e.config.ts
@@ -7,7 +7,7 @@ export default defineConfig({
   test: {
     // Use global to avoid globals imports (describe, test, expect)
     globals: true,
-    include: ['src/**/*.spec.ts'],
+    include: ['src/**/*.e2e-spec.ts'],
     testTimeout: 10 * 60 * 1000, // 10 minutes
   },
 })
